+++ sample-apps/imports-contacts-app/app/lib/services/authorization/get_authorization_uri.rb
@@ -13,7 +13,7 @@ module Services
         ::Hubspot::OAuthHelper.authorize_url(
           client_id: ENV['HUBSPOT_CLIENT_ID'],
           redirect_uri: redirect_uri,
-          scope: %w[contacts]
+          scope: %w[crm.import]
         )
       end