+++ sample-apps/webhooks-contacts-app/db/schema.rb
@@ -12,10 +12,7 @@
 
 ActiveRecord::Schema.define(version: 2020_01_30_113853) do
 
-  # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql"
-
-  create_table "events", force: :cascade do |t|
+  create_table "events", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "event_type"
     t.integer "object_id"
     t.string "event_id"
@@ -26,7 +23,7 @@ ActiveRecord::Schema.define(version: 2020_01_30_113853) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "tokens", force: :cascade do |t|
+  create_table "tokens", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "refresh_token"
     t.string "access_token"
     t.datetime "expires_in"