+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2020-06-01
+
+- Add webhooks contacts sample app
+- Add imports contacts sample app
+- Add webhook signature verification helper
+
 ## [3.2.0] - 2020-05-06
 
 - Rename imports#get_all to get_page method